Split table into deck and pile files, add first animation for throw
