+++ frontend/src/components/game/index.tsx
@@ -12,7 +12,8 @@ import { applyDrag } from './drag';
 
 import { Flash, FlashTransitionTimeout } from '../common/flash';
 import { Scores } from './scores';
-import { Pile, Deck } from './table';
+import { Deck } from './deck';
+import { Pile } from './pile';
 import { Hand } from './hand';
 import { NextGameControls } from './next';
 import { Actions } from './actions';
@@ -452,6 +453,7 @@ export class Game extends Component<
               pile={serverState.currentGame?.pile}
               disabled={!this.isCurrentPlayerDraw()}
               drawAction={this.drawFromPile}
+              lastActionTmp={serverState.currentGame?.lastAction}
             />
             <Deck
               deck={serverState.currentGame?.deck}
